+++ crawl4ai/browser/__init__.py
@@ -6,5 +6,17 @@ for browser creation and interaction.
 
 from .manager import BrowserManager
 from .profiles import BrowserProfileManager
+from .models import DockerConfig
+from .docker_registry import DockerRegistry
+from .docker_utils import DockerUtils
+from .strategies import (
+    BaseBrowserStrategy,
+    PlaywrightBrowserStrategy,
+    CDPBrowserStrategy,
+    BuiltinBrowserStrategy,
+    DockerBrowserStrategy
+)
 
-__all__ = ['BrowserManager', 'BrowserProfileManager']
+__all__ = ['BrowserManager', 'BrowserProfileManager', 'DockerConfig', 'DockerRegistry', 'DockerUtils', 'BaseBrowserStrategy',
+           'PlaywrightBrowserStrategy', 'CDPBrowserStrategy', 'BuiltinBrowserStrategy',
+           'DockerBrowserStrategy']
