+++ crawl4ai/browser/strategies/docker_strategy.py
@@ -6,18 +6,15 @@ which offers better isolation, consistency across platforms, and easy scaling.
 
 import os
 import uuid
-import asyncio
-from typing import Dict, List, Optional, Tuple, Union
-from pathlib import Path
+from typing import List, Optional
 
-from playwright.async_api import Page, BrowserContext
 
-from ..async_logger import AsyncLogger
-from ..async_configs import BrowserConfig, CrawlerRunConfig
-from .docker_config import DockerConfig
-from .docker_registry import DockerRegistry
-from .docker_utils import DockerUtils
-from .strategies import BuiltinBrowserStrategy
+from ...async_logger import AsyncLogger
+from ...async_configs import BrowserConfig
+from ..models import DockerConfig
+from ..docker_registry import DockerRegistry
+from ..docker_utils import DockerUtils
+from .builtin import BuiltinBrowserStrategy
 
 
 class DockerBrowserStrategy(BuiltinBrowserStrategy):
@@ -53,6 +50,16 @@ class DockerBrowserStrategy(BuiltinBrowserStrategy):
         self.docker_config = self.config.docker_config or DockerConfig()
         self.container_id = None
         self.container_name = f"crawl4ai-browser-{uuid.uuid4().hex[:8]}"
+
+        # Use the shared registry file path for consistency with BuiltinBrowserStrategy
+        registry_file = self.docker_config.registry_file
+        if registry_file is None and self.config.user_data_dir:
+            # Use the same registry file as BuiltinBrowserStrategy if possible
+            registry_file = os.path.join(
+                os.path.dirname(self.config.user_data_dir), 
+                "browser_config.json"
+            )
+                    
         self.registry = DockerRegistry(self.docker_config.registry_file)
         self.docker_utils = DockerUtils(logger)
         self.chrome_process_id = None
@@ -60,7 +67,77 @@ class DockerBrowserStrategy(BuiltinBrowserStrategy):
         self.internal_cdp_port = 9222  # Chrome's internal CDP port
         self.internal_mapped_port = 9223  # Port that socat maps to internally
         self.shutting_down = False
-    
+
+    async def start(self):
+        """Start or connect to a browser running in a Docker container.
+        
+        This method initializes Playwright and establishes a connection to 
+        a browser running in a Docker container. Depending on the configured mode:
+        - "connect": Connects to a container with Chrome already running
+        - "launch": Creates a container and launches Chrome within it
+        
+        Returns:
+            self: For method chaining
+        """
+        # Initialize Playwright
+        from ..utils import get_playwright
+        self.playwright = await get_playwright()
+        
+        if self.logger:
+            self.logger.info(
+                f"Starting Docker browser strategy in {self.docker_config.mode} mode",
+                tag="DOCKER"
+            )
+        
+        try:
+            # Get CDP URL by creating or reusing a Docker container
+            # This handles the container management and browser startup
+            cdp_url = await self._get_or_create_cdp_url()
+            
+            if not cdp_url:
+                raise Exception("Failed to establish CDP connection to Docker container")
+            
+            if self.logger:
+                self.logger.info(f"Connecting to browser in Docker via CDP: {cdp_url}", tag="DOCKER")
+            
+            # Connect to the browser using CDP
+            self.browser = await self.playwright.chromium.connect_over_cdp(cdp_url)
+            
+            # Get existing context or create default context
+            contexts = self.browser.contexts
+            if contexts:
+                self.default_context = contexts[0]
+                if self.logger:
+                    self.logger.debug("Using existing browser context", tag="DOCKER")
+            else:
+                if self.logger:
+                    self.logger.debug("Creating new browser context", tag="DOCKER")
+                self.default_context = await self.create_browser_context()
+                await self.setup_context(self.default_context)
+                
+            return self
+        
+        except Exception as e:
+            # Clean up resources if startup fails
+            if self.container_id and not self.docker_config.persistent:
+                if self.logger:
+                    self.logger.warning(
+                        f"Cleaning up container after failed start: {self.container_id[:12]}",
+                        tag="DOCKER"
+                    )
+                await self.docker_utils.remove_container(self.container_id)
+                self.registry.unregister_container(self.container_id)
+                self.container_id = None
+                
+            if self.playwright:
+                await self.playwright.stop()
+                self.playwright = None
+                
+            # Re-raise the exception
+            if self.logger:
+                self.logger.error(f"Failed to start Docker browser: {str(e)}", tag="DOCKER")
+            raise    
+
     async def _generate_config_hash(self) -> str:
         """Generate a hash of the configuration for container matching.
         
@@ -87,7 +164,7 @@ class DockerBrowserStrategy(BuiltinBrowserStrategy):
         # Use the utility method to generate the hash
         return self.docker_utils.generate_config_hash(config_dict)
     
-    async def _get_or_create_cdp_url(self) -> str:
+    async def _get_or_create_cdp_url1(self) -> str:
         """Get CDP URL by either creating a new container or using an existing one.
         
         Returns:
@@ -183,7 +260,109 @@ class DockerBrowserStrategy(BuiltinBrowserStrategy):
         
         # Return CDP URL
         return f"http://localhost:{host_port}"
-    
+
+    async def _get_or_create_cdp_url(self) -> str:
+        """Get CDP URL by either creating a new container or using an existing one.
+        
+        Returns:
+            CDP URL for connecting to the browser
+            
+        Raises:
+            Exception: If container creation or browser launch fails
+        """
+        # If CDP URL is explicitly provided, use it
+        if self.config.cdp_url:
+            return self.config.cdp_url
+
+        # Ensure Docker image exists (will build if needed)
+        image_name = await self.docker_utils.ensure_docker_image_exists(
+            self.docker_config.image, 
+            self.docker_config.mode
+        )
+        
+        # Generate config hash for container matching
+        config_hash = await self._generate_config_hash()
+        
+        # Look for existing container with matching config
+        container_id = await self.registry.find_container_by_config(config_hash, self.docker_utils)
+        
+        if container_id:
+            # Use existing container
+            self.container_id = container_id
+            host_port = self.registry.get_container_host_port(container_id)
+            if self.logger:
+                self.logger.info(f"Using existing Docker container: {container_id[:12]}", tag="DOCKER")
+        else:
+            # Get a port for the new container
+            host_port = self.docker_config.host_port or self.registry.get_next_available_port(self.docker_utils)
+            
+            # Prepare volumes list
+            volumes = list(self.docker_config.volumes)
+            
+            # Add user data directory if specified
+            if self.docker_config.user_data_dir:
+                # Ensure user data directory exists
+                os.makedirs(self.docker_config.user_data_dir, exist_ok=True)
+                volumes.append(f"{self.docker_config.user_data_dir}:{self.docker_config.container_user_data_dir}")
+                
+                # Update config user_data_dir to point to container path
+                self.config.user_data_dir = self.docker_config.container_user_data_dir
+            
+            # Create a new container
+            container_id = await self.docker_utils.create_container(
+                image_name=image_name,
+                host_port=host_port,
+                container_name=self.container_name,
+                volumes=volumes,
+                network=self.docker_config.network,
+                env_vars=self.docker_config.env_vars,
+                cpu_limit=self.docker_config.cpu_limit,
+                memory_limit=self.docker_config.memory_limit,
+                extra_args=self.docker_config.extra_args
+            )
+            
+            if not container_id:
+                raise Exception("Failed to create Docker container")
+            
+            self.container_id = container_id           
+            
+            # Wait for container to be ready
+            await self.docker_utils.wait_for_container_ready(container_id)
+            
+            # Handle specific setup based on mode
+            if self.docker_config.mode == "launch":
+                # In launch mode, we need to start socat and Chrome
+                await self.docker_utils.start_socat_in_container(container_id)
+                
+                # Build browser arguments
+                browser_args = self._build_browser_args()
+                
+                # Launch Chrome
+                await self.docker_utils.launch_chrome_in_container(container_id, browser_args)
+                
+                # Get PIDs for later cleanup
+                self.chrome_process_id = await self.docker_utils.get_process_id_in_container(
+                    container_id, "chrome"
+                )
+                self.socat_process_id = await self.docker_utils.get_process_id_in_container(
+                    container_id, "socat"
+                )
+            
+            # Wait for CDP to be ready
+            cdp_json_config = await self.docker_utils.wait_for_cdp_ready(host_port)
+
+            if cdp_json_config:
+                # Register the container in the shared registry
+                self.registry.register_container(container_id, host_port, config_hash, cdp_json_config)
+            else:
+                raise Exception("Failed to get CDP JSON config from Docker container")
+            
+            if self.logger:
+                self.logger.success(f"Docker container ready: {container_id[:12]} on port {host_port}", tag="DOCKER")
+        
+        # Return CDP URL
+        return f"http://localhost:{host_port}"
+
     def _build_browser_args(self) -> List[str]:
         """Build Chrome command line arguments based on BrowserConfig.
